Address review feedback: fix test quality, reorganize test files
Review fixes:
- Correlation doctest: print exact values [1.0, -1.0] instead of
  threshold check
- test_get_code_set_with_max_by_correlation: replace constant column
  (was producing NaN/ConstantInputWarning) with data that has known
  distinct correlations; assert exact codes_set == {0, 1}
- test_get_code_set_with_min_correlation: same fix, assert exact set
- Remove unnecessary `if: always()` from codecov step
- Collapse 3 copy-pasted xgboost predict split tests into one
  @pytest.mark.parametrize test

File reorganization:
- Delete test_remaining_coverage.py (catch-all "cover the line" file)
- Move TabularDataset tests to test_tabular_dataset.py (12 tests)
- Move model validation tests to test_error_paths.py (12 tests)
- Create test_scripts.py for Hydra script tests (10 tests)
