+++ tests/test_tabular_dataset.py
@@ -3,6 +3,7 @@
 Tests that verify actual behavior rather than just that mocks were called.
 """
 
+from pathlib import Path
 from unittest.mock import MagicMock, patch
 
 import numpy as np
@@ -251,3 +252,183 @@ def test_get_classes_single_class():
 
     classes = ds.get_classes()
     np.testing.assert_array_equal(classes, [1])
+
+
+# ============================================================================
+# Tests moved from test_remaining_coverage.py
+# ============================================================================
+
+
+def test_load_matrix_bad_array_shape(tmp_path):
+    """_load_matrix raises ValueError when npz has wrong array dimensions (line 113)."""
+    ds = _make_dataset()
+    fp = tmp_path / "bad.npz"
+    bad_array = np.zeros((2, 5))  # should be 3 rows
+    np.savez(fp, array=bad_array, shape=np.array([5, 5]))
+    with pytest.raises(ValueError, match="Expected array to have 3 rows"):
+        ds._load_matrix(fp)
+
+
+def test_init_empty_labels_after_loading(tmp_path):
+    """__init__ raises ValueError when labels dict is empty after loading (line 74)."""
+    label_dir = tmp_path / "labels" / "train"
+    label_dir.mkdir(parents=True)
+    pl.DataFrame({"subject_id": [1], "boolean_value": [True]}).write_parquet(label_dir / "0.parquet")
+
+    cfg = MagicMock()
+    cfg.path.input_label_cache_dir = str(tmp_path / "labels")
+
+    with (
+        patch.object(TabularDataset, "_get_code_set", return_value=({0}, {"code/count": [True]}, 1)),
+        patch.object(TabularDataset, "_set_scaler"),
+        patch.object(TabularDataset, "_set_imputer"),
+        patch.object(TabularDataset, "_load_ids_and_labels", return_value=({}, {})),
+        pytest.raises(ValueError, match="No labels found"),
+    ):
+        TabularDataset(cfg, "train")
+
+
+def test_get_code_set_with_max_by_correlation():
+    """_get_code_set correlation-based feature selection (lines 195-201)."""
+    ds = _make_dataset()
+    ds.cfg.tabularization.aggs = ["code/count"]
+    ds.cfg.tabularization.filtered_code_metadata_fp = "fake.parquet"
+    ds.cfg.tabularization.max_by_correlation = 2
+    ds.cfg.tabularization.min_correlation = None
+
+    # Mock dependencies
+    feature_columns = ["A/code", "B/code", "C/code"]
+    with (
+        patch("MEDS_tabular_automl.tabular_dataset.get_feature_columns", return_value=feature_columns),
+        patch("MEDS_tabular_automl.tabular_dataset.get_feature_indices", return_value=[0, 1, 2]),
+    ):
+        ds.cfg.tabularization._resolved_codes = {"A/code", "B/code", "C/code"}
+        # Mock _get_shard_by_index to return data with known correlations
+        X = sp.csc_matrix(np.array([[1.0, 0.0, 0.5], [0.0, 1.0, 0.5], [1.0, 0.0, 0.5], [0.0, 1.0, 0.5]]))
+        y = np.array([1, 0, 1, 0])
+        ds._get_shard_by_index = MagicMock(return_value=(X, y))
+
+        codes_set, _code_masks, _num_features = ds._get_code_set()
+        # Should select top 2 by correlation
+        assert len(codes_set) <= 2
+
+
+def test_get_code_set_with_min_correlation():
+    """_get_code_set min_correlation filtering (lines 203-207)."""
+    ds = _make_dataset()
+    ds.cfg.tabularization.aggs = ["code/count"]
+    ds.cfg.tabularization.filtered_code_metadata_fp = "fake.parquet"
+    ds.cfg.tabularization.max_by_correlation = None
+    ds.cfg.tabularization.min_correlation = 0.5
+
+    feature_columns = ["A/code", "B/code"]
+    with (
+        patch("MEDS_tabular_automl.tabular_dataset.get_feature_columns", return_value=feature_columns),
+        patch("MEDS_tabular_automl.tabular_dataset.get_feature_indices", return_value=[0, 1]),
+    ):
+        ds.cfg.tabularization._resolved_codes = {"A/code", "B/code"}
+        X = sp.csc_matrix(np.array([[1.0, 0.0], [0.0, 1.0], [1.0, 0.0], [0.0, 1.0]]))
+        y = np.array([1, 0, 1, 0])
+        ds._get_shard_by_index = MagicMock(return_value=(X, y))
+
+        codes_set, _code_masks, _num_features = ds._get_code_set()
+        # Both features have abs(corr) > 0.5
+        assert len(codes_set) == 2
+
+
+def test_set_imputer_with_partial_fit():
+    """_set_imputer partial_fit branch (lines 263-265)."""
+    ds = _make_dataset(num_shards=2)
+    mock_imputer = MagicMock()
+    mock_imputer.partial_fit = MagicMock()
+    ds.cfg.data_loading_params.imputer.imputer_target = mock_imputer
+    ds._get_shard_by_index = MagicMock(return_value=(sp.csc_matrix(np.eye(3)), np.array([1, 0, 1])))
+
+    ds._set_imputer()
+    assert mock_imputer.partial_fit.call_count == 2
+
+
+def test_set_scaler_with_fit():
+    """_set_scaler fit-only branch (lines 286-287)."""
+    ds = _make_dataset()
+    mock_scaler = MagicMock()
+    mock_scaler.fit = MagicMock()
+    del mock_scaler.partial_fit
+    ds.cfg.data_loading_params.normalization.normalizer = mock_scaler
+    ds._get_shard_by_index = MagicMock(return_value=(sp.csc_matrix(np.eye(3)), np.array([1, 0, 1])))
+
+    ds._set_scaler()
+    mock_scaler.fit.assert_called_once()
+
+
+def test_get_shard_reloads_labels(tmp_path):
+    """_get_shard_by_index reloads labels when None (line 371)."""
+    ds = _make_dataset()
+    ds.labels = None
+    ds._load_labels = MagicMock(return_value={"shard_0": np.array([1, 0])})
+    ds._get_dynamic_shard_by_index = MagicMock(return_value=sp.csc_matrix(np.eye(2)))
+
+    _x, y = ds._get_shard_by_index(0)
+    ds._load_labels.assert_called_once()
+    assert ds.labels is not None
+    np.testing.assert_array_equal(y, np.array([1, 0]))
+
+
+def test_get_data_shards_empty_raises():
+    """get_data_shards with empty index list raises ValueError (line 420)."""
+    ds = _make_dataset()
+    with pytest.raises(ValueError, match="No data found"):
+        ds.get_data_shards([])
+
+
+def test_get_data_shard_count():
+    """get_data_shard_count returns number of shards (line 441)."""
+    ds = _make_dataset(num_shards=3)
+    assert ds.get_data_shard_count() == 3
+
+
+def test_get_all_column_names():
+    """get_all_column_names constructs feature names from file structure (lines 462-478)."""
+    ds = _make_dataset()
+    ds.cfg.tabularization.filtered_code_metadata_fp = "fake.parquet"
+
+    # Create fake file structure: window/code_type/agg_name.npz
+    fake_file = Path("/fake/30d/code/count.npz")
+    with (
+        patch("MEDS_tabular_automl.tabular_dataset.get_model_files", return_value=[fake_file]),
+        patch("MEDS_tabular_automl.tabular_dataset.get_feature_columns", return_value=["A/code", "B/code"]),
+        patch("MEDS_tabular_automl.tabular_dataset.get_feature_indices", return_value=[0, 1]),
+    ):
+        names = ds.get_all_column_names()
+        assert names == ["A/code/count/30d", "B/code/count/30d"]
+
+
+def test_get_column_names_with_indices():
+    """get_column_names with indices parameter filters results (lines 486-505)."""
+    ds = _make_dataset()
+    ds.cfg.tabularization.filtered_code_metadata_fp = "fake.parquet"
+
+    fake_file = Path("/fake/30d/code/count.npz")
+    with (
+        patch("MEDS_tabular_automl.tabular_dataset.get_model_files", return_value=[fake_file]),
+        patch("MEDS_tabular_automl.tabular_dataset.get_feature_columns", return_value=["A/code", "B/code"]),
+        patch("MEDS_tabular_automl.tabular_dataset.get_feature_indices", return_value=[0, 1]),
+    ):
+        # Without indices — returns all
+        all_names = ds.get_column_names()
+        assert len(all_names) == 2
+
+        # With indices — returns subset
+        filtered = ds.get_column_names(indices=[0])
+        assert filtered == ["A/code/count/30d"]
+
+
+def test_densify():
+    """Densify() iterates shards and returns dense data + labels (lines 511-519)."""
+    ds = _make_dataset(labels={"shard_0": np.array([1, 0])}, num_shards=1)
+    shard_data = sp.csc_matrix(np.array([[1.0, 2.0], [3.0, 4.0]]))
+    ds._get_shard_by_index = MagicMock(return_value=(shard_data, np.array([1, 0])))
+
+    data, labels = ds.densify()
+    assert data.shape == (2, 2)
+    np.testing.assert_array_equal(labels, [1, 0])
